requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.13.2
    # via edge-tts
aiosignal==1.4.0
    # via aiohttp
annotated-types==0.7.0
    # via pydantic
anyio==4.11.0
    # via httpx
attrs==25.4.0
    # via aiohttp
certifi==2025.11.12
    # via
    #   edge-tts
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.4
    # via requests
colorama==0.4.6
    # via tqdm
decorator==5.2.1
    # via moviepy
edge-tts==7.2.3
    # via langgraph-automator (pyproject.toml)
frozenlist==1.8.0
    # via
    #   aiohttp
    #   aiosignal
h11==0.16.0
    # via httpcore
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via
    #   langgraph-sdk
    #   langsmith
    #   ollama
idna==3.11
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
imageio==2.37.2
    # via moviepy
imageio-ffmpeg==0.6.0
    # via moviepy
jsonpatch==1.33
    # via langchain-core
jsonpointer==3.0.0
    # via jsonpatch
langchain-core==1.1.0
    # via
    #   langgraph-automator (pyproject.toml)
    #   langchain-ollama
    #   langgraph
    #   langgraph-checkpoint
    #   langgraph-prebuilt
langchain-ollama==1.0.0
    # via langgraph-automator (pyproject.toml)
langgraph==1.0.4
    # via langgraph-automator (pyproject.toml)
langgraph-checkpoint==3.0.1
    # via
    #   langgraph
    #   langgraph-prebuilt
langgraph-prebuilt==1.0.5
    # via langgraph
langgraph-sdk==0.2.10
    # via langgraph
langsmith==0.4.49
    # via langchain-core
moviepy==2.2.1
    # via langgraph-automator (pyproject.toml)
multidict==6.7.0
    # via
    #   aiohttp
    #   yarl
numpy==2.3.5
    # via
    #   imageio
    #   moviepy
ollama==0.6.1
    # via langchain-ollama
orjson==3.11.4
    # via
    #   langgraph-sdk
    #   langsmith
ormsgpack==1.12.0
    # via langgraph-checkpoint
packaging==25.0
    # via
    #   langchain-core
    #   langsmith
pillow==11.3.0
    # via
    #   langgraph-automator (pyproject.toml)
    #   imageio
    #   moviepy
proglog==0.1.12
    # via moviepy
propcache==0.4.1
    # via
    #   aiohttp
    #   yarl
pydantic==2.12.5
    # via
    #   langgraph-automator (pyproject.toml)
    #   langchain-core
    #   langgraph
    #   langsmith
    #   ollama
pydantic-core==2.41.5
    # via pydantic
python-dotenv==1.2.1
    # via moviepy
pyyaml==6.0.3
    # via langchain-core
requests==2.32.5
    # via
    #   langgraph-automator (pyproject.toml)
    #   langsmith
    #   requests-toolbelt
requests-toolbelt==1.0.0
    # via langsmith
sniffio==1.3.1
    # via anyio
tabulate==0.9.0
    # via edge-tts
tenacity==9.1.2
    # via langchain-core
tqdm==4.67.1
    # via proglog
typing-extensions==4.15.0
    # via
    #   edge-tts
    #   langchain-core
    #   pydantic
    #   pydantic-core
    #   typing-inspection
typing-inspection==0.4.2
    # via pydantic
urllib3==2.5.0
    # via requests
websocket-client==1.9.0
    # via langgraph-automator (pyproject.toml)
xxhash==3.6.0
    # via langgraph
yarl==1.22.0
    # via aiohttp
zstandard==0.25.0
    # via langsmith
